Resolved comments Amijs
