+++ src/components/filter/Filter.tsx
@@ -7,16 +7,13 @@ interface FilterProps {
 }
 
 export const Filter = ({ setCurrentClient }: FilterProps) => {
-  const { teamEntries, setTeamEntries } = useContext(TeamEntriesContext);
+  const { teamEntries } = useContext(TeamEntriesContext);
 
   const handleFilter = (event: React.ChangeEvent<HTMLSelectElement>) => {
     setCurrentClient(event.target.value);
   };
 
-  const onlyClients = teamEntries.map(function (entry) {
-    return entry.client;
-  });
-  const formattedClients = [...new Set(onlyClients)];
+  const clients = [...new Set(teamEntries.map((entry) => entry.client))];
 
   return (
     <>
@@ -24,7 +21,7 @@ export const Filter = ({ setCurrentClient }: FilterProps) => {
         <Styled.Label>Filter by client:</Styled.Label>
         <Styled.Select name="clients" id="client-filter" onChange={handleFilter}>
           <option value="">Choose a client</option>
-          {formattedClients.map((entry) => {
+          {clients.map((entry) => {
             return (
               <option value={entry} label={entry} key={entry}>
                 {entry}
