+++ src/components/team-entries/TeamEntries.tsx
@@ -13,38 +13,36 @@ interface TeamEntriesProps {
 }
 
 export const TeamEntries = ({ isModalActive, handleModal, currentClient }: TeamEntriesProps) => {
-  const { teamEntries, setTeamEntries } = useContext(TeamEntriesContext);
+  const { teamEntries } = useContext(TeamEntriesContext);
 
   return (
-    <>
-      <Styled.Container>
-        <Modal isActive={isModalActive} onClose={handleModal} title={"New Humanoid"}>
-          <TeamEntryForm isActive={isModalActive} handleModal={handleModal} />
-        </Modal>
-        {teamEntries
-          .filter((teamEntry) => currentClient === "" || teamEntry.client === currentClient)
-          .map((teamEntry, i, arr) => {
-            const currentClient = teamEntry.client;
-            const previousClient = arr[i - 1]?.client;
-            return (
-              <div key={Math.random()}>
-                {previousClient !== currentClient && (
-                  <Styled.Section>
-                    <Styled.Client>{currentClient}</Styled.Client>
-                  </Styled.Section>
-                )}
-                <TeamEntry
-                  client={teamEntry.client}
-                  firstName={teamEntry.firstName}
-                  lastName={teamEntry.lastName}
-                  role={teamEntry.role}
-                  startDate="February 2022"
-                  key={teamEntry.id}
-                />
-              </div>
-            );
-          })}
-      </Styled.Container>
-    </>
+    <Styled.Container>
+      <Modal isActive={isModalActive} onClose={handleModal} title={"New Humanoid"}>
+        <TeamEntryForm isActive={isModalActive} handleModal={handleModal} />
+      </Modal>
+      {teamEntries
+        .filter((teamEntry) => currentClient === "" || teamEntry.client === currentClient)
+        .map((teamEntry, i, arr) => {
+          const currentClient = teamEntry.client;
+          const previousClient = arr[i - 1]?.client;
+          return (
+            <div key={Math.random()}>
+              {previousClient !== currentClient && (
+                <Styled.Section>
+                  <Styled.Client>{currentClient}</Styled.Client>
+                </Styled.Section>
+              )}
+              <TeamEntry
+                client={teamEntry.client}
+                firstName={teamEntry.firstName}
+                lastName={teamEntry.lastName}
+                role={teamEntry.role}
+                startDate="February 2022"
+                key={teamEntry.id}
+              />
+            </div>
+          );
+        })}
+    </Styled.Container>
   );
 };
